Refactor : 파일 위치 변경, 주석 삭제, 경로 변경
